requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
py3-tts==3.5.0
blobfile==2.1.1
openai==1.2.3
openai-whisper @ git+https://github.com/openai/whisper.git@fcfeaf1b61994c071bba62da47d7846933576ac9
PyAudio==0.2.14
pyyaml==6.0.1
pygame==2.5.2